First Commit
ABC Kids App is my first personal project to be deployed to the Google
App Store.  This application will help children build familiarity with
the English Alphabet.  After tapping "New Game" from the main page,
users will be faced with a random image of an object.  On the bottom
portion of the screen, the user will be able to select from four letters
(3 incorrect and 1 correct).  The letter must match the first letter of
the object.  As time goes on, I may add more features, but the ultimate
goal is to build a project from start to finish and have it ready on the
Google Play Store.